Split rules conformance corpus into one pack record per file
Replace the aggregate pack files (stress-packs.ts, fix-class-packs.ts,
storage/packs.ts) with one authored record per file, named <pack-id>.ts,
per the ratified data-record convention. The filename is the pack id;
authored records carry no id field. A loader per corpus directory
(load.ts, mirroring rigs/load.ts and surfaces/load.ts) reads the
directory, validates each record (unique case descriptions, legal
ALLOW/DENY expectations, non-empty rules source, filename-safe id,
no stray id field), and computes ALL_RULES_FIRESTORE_PACKS /
ALL_RULES_STORAGE_PACKS.

The Firestore corpus also carries a loader-only `group` field
('stress' | 'fix-class') so index.ts can still compute STRESS_PACKS and
FIX_CLASS_PACKS for the two live-parity suites that consume them
separately; `group` is stripped before any Pack reaches another
consumer, so the Pack/StoragePack shape every existing consumer reads
(capture runners, replay suites, live parity harness) is unchanged.

Pack literals were extracted with the TypeScript compiler API (not
regex) to keep rules/cases byte-identical; a throwaway verification
script confirmed zero content mismatches across all 25 packs before the
aggregate files were deleted.
