+++ packages/conformance/rules-corpus/firestore/types.ts
@@ -33,4 +33,26 @@ export interface Pack {
   cases: TestCase[];
 }
 
+/**
+ * Which historical aggregate a pack came from: the resurrected pre-cutover
+ * suite (`stress`, formerly stress-packs.ts) or a round-1/2 remediation
+ * (`fix-class`, formerly fix-class-packs.ts). The live-parity suites
+ * (packages/pyric/test/rules/parity/parity-stress.test.ts and
+ * round-fix-classes.test.ts) run these as two separate reports, so the
+ * distinction is authored per pack rather than inferred.
+ */
+export type PackGroup = 'stress' | 'fix-class';
+
+/**
+ * The authored shape for one file in rules-corpus/firestore/. The filename
+ * IS the pack id, so the record carries no `id` field — the loader
+ * (./load.ts) injects it from the filename. `group` is loader-only
+ * classification metadata: it is consumed to reconstruct STRESS_PACKS /
+ * FIX_CLASS_PACKS and is stripped before a `Pack` reaches any other
+ * consumer, so the `Pack` type itself (and everything that already reads
+ * it — the capture runner, the replay suite, the live parity harness)
+ * stays unchanged.
+ */
+export type PackRecord = Omit<Pack, 'id'> & { group: PackGroup };
+
 export type { TestCase };
